// app/schemas/blog.ts
import z from "zod";

export const blogSchema = z.object({
    title: z.string().min(3, "Title must be at least 3 characters long").max(100, "Title must be at most 100 characters long"),
    content: z.string().min(10, "Content must be at least 10 characters long").max(10000, "Content must be at most 10000 characters long"),
    image: z.instanceof(File),   
})